+++ analytic_plan/__init__.py
@@ -1,11 +1,10 @@
 # -*- coding: utf-8 -*-
-# © 2015 Eficent Business and IT Consulting Services S.L.
+# Copyright 2015 Eficent Business and IT Consulting Services S.L.
 # (Jordi Ballester Alomar)
-#
-# © 2016 Matmoz d.o.o.
+# Copyright 2016 Matmoz d.o.o.
 # (Matjaž Mozetič)
-#
+# Copyright 2017 Serpent Consulting Services Pvt. Ltd.
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-import models
-import wizard
+from . import models
+from . import wizard
