+++ analytic_plan/models/__init__.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
-# © 2015 Eficent Business and IT Consulting Services S.L.
+# Copyright 2015 Eficent Business and IT Consulting Services S.L.
 # (Jordi Ballester Alomar)
-#
-# © 2016 Matmoz d.o.o.
+# Copyright 2016 Matmoz d.o.o.
 # (Matjaž Mozetič)
-#
+# Copyright 2017 Serpent Consulting Services Pvt. Ltd.
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 from . import account_analytic_plan_version
+from . import account_analytic_plan
 from . import account_analytic_plan_journal
 from . import account_analytic_account
