+++ analytic_plan/models/account_analytic_plan_journal.py
@@ -1,13 +1,12 @@
 # -*- coding: utf-8 -*-
-# © 2015 Eficent Business and IT Consulting Services S.L.
+# Copyright 2015 Eficent Business and IT Consulting Services S.L.
 # (Jordi Ballester Alomar)
-#
-# © 2016 Matmoz d.o.o.
+# Copyright 2016 Matmoz d.o.o.
 # (Matjaž Mozetič)
-#
+# Copyright 2017 Serpent Consulting Services Pvt. Ltd.
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from openerp import api, fields, models
+from odoo import api, fields, models
 
 
 class AccountAnalyticPlanJournal(models.Model):
@@ -16,33 +15,31 @@ class AccountAnalyticPlanJournal(models.Model):
     _description = 'Analytic Journal Plan'
 
     name = fields.Char(
-        'Planning Journal Name', required=True
+        'Planning Journal Name',
+        required=True
     )
     code = fields.Char(
         'Planning Journal Code'
     )
     active = fields.Boolean(
         'Active',
-        help="If the active field is set to False, "
-             "it will allow you to hide the analytic "
-             "journal without removing it.",
+        help="""If the active field is set to False, it will allow you to
+                hide the analytic journal without removing it.""",
         default=True
     )
     type = fields.Selection(
         [
-            ('sale', 'Sale'),
-            ('purchase', 'Purchase'),
-            ('cash', 'Cash'),
-            ('general', 'General'),
-            ('situation', 'Situation')
+         ('sale', 'Sale'),
+         ('purchase', 'Purchase'),
+         ('cash', 'Cash'),
+         ('general', 'General'),
+         ('situation', 'Situation')
         ],
         'Type',
         required=True,
-        help="Gives the type of the analytic "
-             "journal. When it needs for a document "
-             "(eg: an invoice) to create analytic "
-             "entries, OpenERP will look  for a "
-             "matching journal of the same type.",
+        help="""Gives the type of the analytic journal. When it needs for a
+                document (eg: an invoice) to create analytic entries, OpenERP
+                will look  for a matching journal of the same type.""",
         default='general'
     )
     line_ids = fields.One2many(
@@ -57,7 +54,7 @@ class AccountAnalyticPlanJournal(models.Model):
         default=lambda self:
             self.env['res.users'].browse(self._uid).company_id.id)
     analytic_journal = fields.Many2one(
-        'account.analytic.journal',
-        'Actual Analytic journal',
+        'account.journal',
+        'Actual Journal',
         required=False
     )
