+++ analytic_plan/static/description/index.html
@@ -6,60 +6,54 @@
         </div>
         <div class="oe_span6">
             <p class="oe_mt32">
-            <a href="http://www.eficent.com"> Eficent</a> brings you this module to better manage the planned costs of your projects or analytic accounts.
-			</p>             
+                <a href="http://www.eficent.com"> Eficent</a> brings you this module to better manage the planned costs of your projects or analytic accounts.
+            </p>
         </div>
     </div>
 </section>
 
 <section class="oe_container oe_dark">
-      <h2 class="oe_slogan">New planning journal</h2>
-      <div class="oe_span6">
-		<p class="oe_mt32">
-			A planning journal is created. It is similar to the analytic journal, but used for planning purposes.
-		</p>
-		<p class="oe_mt32">
-			The planning analytic journals can be configured. 
-		</p>			
-		<p class="oe_mt32">
-			A planning analytic journal lines object is created, with the exceptions of referencing to the planning analytic journal instead of the analytic journal, and considering  that the general account is not a required entry.  
-		</p>
-		<p class="oe_mt32">
-			The new object is visible as a separate entity, accesible from the Accounting area, with the corresponding search, tree, form views.  
-		</p>
-		<p class="oe_mt32">
-			New analytic account planning attributes: cumulated planned costs, cummulated planned earnings and cumulated balance. The attributes are calculated based on the planning analytic journal lines.  
-		</p>
-		<p class="oe_mt32">
-			The new attributes are visible on the following views: Analytic account forms: cumulated planned costs, cumulated planned earnings and cumulated balance; Budget positions: cumulated planned costs  
-		</p>
-		<p class="oe_mt32">
-			A report exists to create a PDF with the corresponding journal entries.
-		</p>
-	</div>
-      <h2 class="oe_slogan">Planning versions</h2>
-      <div class="oe_span6">
-		<p class="oe_mt32">
-			A different planning versions can be maintained.
-            A default version for cost planning is defined,
-            that is proposed every time you enter an analtic planning entry.
-            You can flag one of the versions as the default for commited costs.
-		</p>
-		<p class="oe_mt32">
-			The planning analytic versions can be configured.
-		</p>
-		<p class="oe_mt32">
-			A planning analytic version object is created.
-		</p>
-		<p class="oe_mt32">
-			The new object is visible as a separate entity,
-            accesible from the Accounting area,
-            with the corresponding search, tree, form views.
-		</p>
-	</div>
-  	<h2 class="oe_slogan">More information and assistance</h2>
-    <div class="oe_span6">
-        <p class="oe_mt32">If you are interested in this module and seek further assistance to use it please visit us at <a href="http://www.eficent.com">www.eficent.com</a> or conact us at contact@eficent.com.
-        </p>
-    </div>    
+    <div class="oe_row oe_spaced">
+        <h2 class="oe_slogan">New planning journal</h2>
+        <div class="oe_span12">
+            <p class="oe_mt32" align="justify">
+                A planning journal is created. It is similar to the analytic journal, but used for planning purposes.
+            </p>
+            <p class="oe_mt32">
+                The planning analytic journals can be configured. 
+            </p>            
+            <p class="oe_mt32">
+                A planning analytic journal lines object is created, with the exceptions of referencing to the planning analytic journal instead of the analytic journal, and considering  that the general account is not a required entry.  
+            </p>
+            <p class="oe_mt32">
+                The new object is visible as a separate entity, accesible from the Accounting area, with the corresponding search, tree, form views.  
+            </p>
+            <p class="oe_mt32">
+                New analytic account planning attributes: cumulated planned costs, cummulated planned earnings and cumulated balance. The attributes are calculated based on the planning analytic journal lines.  
+            </p>
+            <p class="oe_mt32">
+                The new attributes are visible on the following views: Analytic account forms: cumulated planned costs, cumulated planned earnings and cumulated balance; Budget positions: cumulated planned costs  
+            </p>
+            <p class="oe_mt32">
+                A report exists to create a PDF with the corresponding journal entries.
+            </p>
+            <h2 class="oe_slogan">Planning versions</h2>
+            <p class="oe_mt32">
+                A different planning versions can be maintained.A default version for cost planning is defined,that is proposed every time you enter an analtic planning entry.You can flag one of the versions as the default for commited costs.
+            </p>
+            <p class="oe_mt32">
+                The planning analytic versions can be configured.
+            </p>
+            <p class="oe_mt32">
+                A planning analytic version object is created.
+            </p>
+            <p class="oe_mt32">
+                The new object is visible as a separate entity,accesible from the Accounting area,with the corresponding search, tree, form views.
+            </p>
+            <h2 class="oe_slogan">More information and assistance</h2>
+            <p class="oe_mt32">
+               If you are interested in this module and seek further assistance to use it please visit us at <a href="http://www.eficent.com">www.eficent.com</a> or conact us at contact@eficent.com.
+            </p>
+        </div>
+    </div>
 </section>
