+++ analytic_plan/views/account_analytic_account_view.xml
@@ -1,82 +1,77 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
+<odoo>
 
-
-		<record id="view_account_analytic_account_tree_inh" model="ir.ui.view">
-            <field name="name">account.analytic.account.tree</field>
-            <field name="model">account.analytic.account</field>
-            <field name="type">tree</field>
-            <field name="field_parent">child_complete_ids</field>
-            <field name="inherit_id"
-                   ref="analytic.view_account_analytic_account_list"/>
-            <field name="arch" type="xml">         	   
-            	<field name="balance" position="after">    
-            	    <field name="debit_plan"/>            	   
-                    <field name="credit_plan"/>                    
-                    <field name="balance_plan"/>
-					<button string="Planned cost lines"
+    <record id="view_account_analytic_account_tree_inh" model="ir.ui.view">
+        <field name="name">account.analytic.account.tree</field>
+        <field name="model">account.analytic.account</field>
+        <field name="type">tree</field>
+        <field name="field_parent">child_complete_ids</field>
+        <field name="inherit_id" ref="analytic.view_account_analytic_account_list"/>
+        <field name="arch" type="xml">
+            <field name="balance" position="after">
+                <field name="debit_plan"/>
+                <field name="credit_plan"/>
+                <field name="balance_plan"/>
+                <button string="Planned cost lines"
                         name="open_plan_cost_tree_view"
                         type="object"
                         icon="gtk-open"/>
-            	</field>
-                <field name="debit" position="before">
-                    <field name="active_analytic_planning_version" string="Active Plan Version"/>
-                </field>
             </field>
-        </record>
-
-
-        <record id="view_account_analytic_account_form_inh" model="ir.ui.view">
-            <field name="name">account.analytic.account.form</field>
-            <field name="model">account.analytic.account</field>
-            <field name="type">form</field>
-            <field name="inherit_id" ref="analytic.view_account_analytic_account_form"/>
-            <field name="arch" type="xml">
-            	<field name="code" position="after">
-            		<field name="active_analytic_planning_version" string="Active Plan Version"/>
-            	</field>
+            <field name="debit" position="before">
+                <field name="active_analytic_planning_version" string="Active Plan Version"/>
             </field>
-        </record>
+        </field>
+    </record>
 
-        <record model="ir.actions.act_window" id="act_analytic_plan_cost_revenue">            
-            <field name="domain">[('account_id','child_of', active_id)]</field>
-            <field name="name">Planned Costs &amp; Revenues</field>
-            <field name="res_model">account.analytic.line.plan</field>
-            <field name="src_model">account.analytic.account</field>
-            <field name="view_mode">tree,form</field>
-            <field name="view_type">form</field>
-            <field name="help" type="html">
-              <p>
-                No activity yet on this contract.
-              </p><p>
-                In OpenERP, contracts and projects are implemented using
-                analytic account. So, you can track costs and revenues to analyse
-                your margins easily.
-              </p>
+    <record id="view_account_analytic_account_form_inh" model="ir.ui.view">
+        <field name="name">account.analytic.account.form</field>
+        <field name="model">account.analytic.account</field>
+        <field name="type">form</field>
+        <field name="inherit_id" ref="analytic.view_account_analytic_account_form"/>
+        <field name="arch" type="xml">
+            <field name="code" position="after">
+                <field name="active_analytic_planning_version" string="Active Plan Version"/>
             </field>
-        </record>
+        </field>
+    </record>
 
-        <record id="view_account_analytic_account_form_analytic_plan" model="ir.ui.view">
-            <field name="name">account.analytic.account.form.proj.cost</field>
-            <field name="model">account.analytic.account</field>
-            <field name="inherit_id"
-                   ref="analytic.view_account_analytic_account_form"/>
-            <field name="arch" type="xml">
-                <xpath expr='//div[@name="button_box"]' position='inside'>
-                    <button help="Planned costs and revenues"
+    <record model="ir.actions.act_window" id="act_analytic_plan_cost_revenue">
+        <field name="domain">[('account_id','child_of', active_id)]</field>
+        <field name="name">Planned Costs &amp; Revenues</field>
+        <field name="res_model">account.analytic.line.plan</field>
+        <field name="src_model">account.analytic.account</field>
+        <field name="view_mode">tree,form</field>
+        <field name="view_type">form</field>
+        <field name="help" type="html">
+          <p>
+            No activity yet on this contract.
+          </p><p>
+            In OpenERP, contracts and projects are implemented using
+            analytic account. So, you can track costs and revenues to analyse
+            your margins easily.
+          </p>
+        </field>
+    </record>
+
+    <record id="view_account_analytic_account_form_analytic_plan" model="ir.ui.view">
+        <field name="name">account.analytic.account.form.proj.cost</field>
+        <field name="model">account.analytic.account</field>
+        <field name="inherit_id"
+               ref="analytic.view_account_analytic_account_form"/>
+        <field name="arch" type="xml">
+            <xpath expr='//div[@name="button_box"]' position='inside'>
+                <button help="Planned costs and revenues"
                         name="%(act_analytic_plan_cost_revenue)d"
                         type="action"
                         context="{'default_account_id': active_id,
                         'search_default_account_id': active_id,
                         'default_version_id': active_analytic_planning_version}"
                         groups="analytic.group_analytic_accounting"
-						class="oe_inline oe_stat_button" icon="fa-pencil-square-o">
-						<div>Planned costs<br/>and revenues</div>
-					</button>
-                </xpath>
-            </field>
-        </record>
+                        class="oe_inline oe_stat_button" icon="fa-pencil-square-o">
+                    <span>Planned costs and revenues</span>
+                </button>
+            </xpath>
+        </field>
+    </record>
 
-    </data>
-</openerp>
+</odoo>
