+++ analytic_plan/views/product_view.xml
@@ -1,18 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <record id="product_template_form_view" model="ir.ui.view">
-            <field name="name">product.template.form.inherit</field>
-            <field name="model">product.template</field>
-            <field name="priority">5</field>
-            <field name="inherit_id" ref="account.product_template_form_view"/>
-            <field name="arch" type="xml">
-                <xpath
-                        expr='//group[@name="properties"]'
-                       position='after'>
-                    <group name="plan_journals" colspan="3"/>
-                </xpath>
-            </field>
-        </record>
-    </data>
-</openerp>
+<odoo>
+
+    <record id="product_template_form_view" model="ir.ui.view">
+        <field name="name">product.template.form.inherit</field>
+        <field name="model">product.template</field>
+        <field name="priority">5</field>
+        <field name="inherit_id" ref="account.product_template_form_view"/>
+        <field name="arch" type="xml">
+            <xpath expr='//group[@name="properties"]'
+                   position='after'>
+                <group name="plan_journals" colspan="3"/>
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
