+++ analytic_plan/views/project_view.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <record id="edit_project_inh" model="ir.ui.view">
-            <field name="name">project.project.form</field>
-            <field name="model">project.project</field>
-            <field name="type">form</field>
-            <field name="inherit_id" ref="project_wbs.edit_project"/>
-            <field name="arch" type="xml">
-            	<xpath expr='//field[@name="project_analytic_account_id"]'
-                       position='after'>
-					<field name="active_analytic_planning_version" string="Active Plan Version"/>
-            	</xpath>
-            </field>
-        </record>
-    </data>
-</openerp>
+<odoo>
+
+    <record id="edit_project_inh" model="ir.ui.view">
+        <field name="name">project.project.form</field>
+        <field name="model">project.project</field>
+        <field name="type">form</field>
+        <field name="inherit_id" ref="project_wbs.edit_project"/>
+        <field name="arch" type="xml">
+            <xpath expr='//field[@name="project_analytic_id"]'
+                   position='after'>
+                <field name="active_analytic_planning_version" string="Active Plan Version"/>
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
