+++ analytic_plan/wizard/__init__.py
@@ -1,13 +1,9 @@
 # -*- coding: utf-8 -*-
-# © 2015 Eficent Business and IT Consulting Services S.L.
+# Copyright 2015 Eficent Business and IT Consulting Services S.L.
 # (Jordi Ballester Alomar)
-#
-# © 2015 Serpent Consulting Services Pvt. Ltd.
-# (Sudhir Arya)
-#
-# © 2016 Matmoz d.o.o.
+# Copyright 2017 Serpent Consulting Services Pvt. Ltd.
+# Copyright 2016 Matmoz d.o.o.
 # (Matjaž Mozetič)
-#
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-import analytic_plan_copy_version
+from . import analytic_plan_copy_version
